+++ p2f_pydantic/harm_data_record.py
@@ -1,7 +1,8 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from typing import Optional
 from uuid import UUID
 
 class HARM_Data_Record(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
     fk_dataset: UUID
     record_hash: str