+++ p2f_pydantic/harm_data_types.py
@@ -1,8 +1,9 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from typing import Optional
 from uuid import UUID
 
 class HARM_Data_Type(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
     datatype_id: Optional[UUID] = None
     measure: str
     unit_of_measurement: str