+++ p2f_pydantic/harm_reference.py
@@ -1,9 +1,10 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from pydantic import HttpUrl
 from typing import Optional
 from uuid import UUID
 
 class HARM_Reference(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
     reference_id: Optional[UUID] = None
     # Now a string because storing the 10prefix/suffix is probably better
     doi: Optional[str] = None