+++ p2f_pydantic/harm_species.py
@@ -1,8 +1,9 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from typing import Optional
 from uuid import UUID
 
 class HARM_Species(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
     species_id: Optional[UUID] = None
     display_species: str
     common_name: Optional[str] = None