+++ p2f_pydantic/harm_timeslices.py
@@ -1,8 +1,9 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from typing import Optional
 from uuid import UUID
 
 class HARM_Timeslice(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
     timeslice_id: Optional[UUID] = None
     timeslice_name: str
     timeslice_age_mean: int