+++ p2f_pydantic/link_git.py
@@ -1,10 +1,11 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from pydantic import AnyUrl
 from datetime import date
 from typing import Optional
 from uuid import UUID
 
 class Git_Repository(BaseModel):
+    model_config = ConfigDict(from_attributes=True)
     git_repo_id: Optional[UUID] = None
     git_repo_url: AnyUrl
     is_p2f_repo: bool = False